Refactore the youtube module
First the method `Get_Data_Details` should return the data instead add it to the database. It's role of the user of this function to decide what to do with the data.

Besides, raise an exception when the playlist has entries with zero elements. With it, we can avoid repeating the code that checks if it has zero elements.